Used a value converter to get rid of the dodgy width stuff made all cell widths in the same column consistent by recording a max width on each collumn header.
Next step:
Calculate the width better. The current calc only works with square fonts and relies on a guessed number in xaml
I won't bother measuring labels but will use a Math.Clamp, to give a minimum and maximum value. This requires net standard 2.1 or I have to right it myself