Upgraded to NetStandard 2.1 for both of the NetStandard Projects.
I may regret this as I think people will want to reference this from 2.0 projects which is the current default for new forms solutions. I don't fully understand the dependancy chain here, maybe it's fine.

If This becomes an issue, the only known feature of 2.1 that I want to use is Math.Clamp, which I can easily implement myself